Make ultratrace into a python module
If we're going to be splitting up functionality into multiple files,
we'll need to have a module structure[1].  That requires our source
files to all live in one directory containing an __init__.py file.

This patch also renames UltraTrace.py to ultratrace/main.py, which is
still the entrypoint for the program.

[1] https://docs.python.org/3/tutorial/modules.html