Add setup.py
Added an install script and a setup.py to faciliate the installation
process. The install script is very rudimentary; it tries to use
Homebrew to install 'portaudio' and gives up if Homebrew is not
installed.

The setup.py file will allow the user to install using `pip`. To build
from source, the user first runs `install.sh` then `pip install .`.
Optionally, the user can pass `--editable` to `pip install` so that
changes to .py files do not require another invocation of install to
have the most up-to-date verison locally.