download-test-data: Split `hash-path` into its own script
Since we'll be moving to a symlink-ish system for storing test data
blobs, we'll want to have consistent hashing across different scripts.

Also, this allows us to be a bit more exacting in our attention to OS
interoperability.  Specifically, it fixes a bunch of issues that were
causing MacOS and Ubuntu to hash the same directory trees to different
values (and thus triggering excessive downloads).