FileLoaderBase: remove the `path = property(...)` line
This was causing issues for the class hierarchy: child classes were
not able to be instantiated because they were getting the error `Can't
instantiate abstract class with abstract methods: "path"`.  To work
around this, we can just require all references to `FileLoaderBase#path`
to be accessed via `::[gs]et_path()`.