AlignmentFileLoader: Add an abstract ::get_intervals() method
This method should return a Sequence of (name, Sequence[Interval])
tuples, where each interval corresponds to one "drawable" alignment
interval.

The first axis corresponds to what we would think of as "vertical"
differences -- the different tiers in TextGridWorld -- and the second
axis corresponds to "horizontal" -- the different intervals of a
specific tier in TextGridWorld.  The reason we iterate Tuples instead
of just using, say, an OrderedDict, is because the tier names can have
duplicates.

This patch also implements some tests :^)