TextGridLoader: Correctly parse ASCII, Latin-1, Mac-Roman encodings
This patch also includes some example textgrid files adapted from the
Praat man page on File Formats[1].  According to that page, the textgrid
library we're using does not completely follow the spec (it assumes we
are always giving it human-readable input).  However, it does detail
some "expected" encodings, so those are implemented in this patch.

[1] http://www.fon.hum.uva.nl/praat/manual/TextGrid_file_formats.html